refactor(list-examples): adjust folder structure of examples
