+++ src/app/examples/list/components/bold-text-on-row-selection.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { BaseListComponent } from '../base-list.component';
+import { BaseListComponent } from './base-list.component';
 
 export const ListBoldTextOnRowSelectionExampleTemplate = `<kirby-list [items]="items" [markSelectedRow]="true">
   <kirby-list-item
