+++ src/app/examples/list/components/colored-items.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { BaseListComponent } from '../base-list.component';
+import { BaseListComponent } from './base-list.component';
 
 export const ListColoredItemsExampleTemplate = `<kirby-list
   [items]="items"
