+++ src/app/examples/list/components/default.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { BaseListComponent } from '../base-list.component';
+import { BaseListComponent } from './base-list.component';
 
 export const ListDefaultExampleTemplate = `<kirby-list [items]="items">
   <kirby-list-item *kirbyListItem="let item" [title]="item.title" [amount]="item.amount">
