+++ src/app/examples/list/components/no-shape/list-no-shape-example.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../kirby/scss/utils';
+@import '../../../../../kirby/scss/utils';
 
 kirby-card {
   margin: get-shadow-size(2);
