+++ src/app/examples/list/components/no-shape/list-no-shape-example.component.ts
