+++ src/app/examples/list/components/swipe/list-swipe-example.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../kirby/scss/utils';
+@import '../../../../../kirby/scss/utils';
 
 $item-height: size('xxxl');
 $dot-size: size('xxs');
