+++ src/app/examples/list/components/swipe/list-swipe-example.component.ts
